requirement.txt
arrow==1.1.1
asgiref==3.4.1
blessed==1.18.1
boto3==1.21.26
botocore==1.24.26
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.5
django-cors-headers==3.7.0
django-picklefield==3.0.1
django-q==1.3.9
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docopt==0.4.0
ecdsa==0.17.0
idna==3.3
jmespath==1.0.0
mandrill==1.0.60
Pillow==9.0.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
PyJWT==2.1.0
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2021.1
redis==3.5.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.9
wcwidth==0.2.5
